Move the create application function to own file
This separates it from the application class itself.
#922
